refactor after review. continue. 2nd attempt
