+++ src/utils/render.js
@@ -9,12 +9,12 @@ export function createDOMElement(content) {
 }
 
 export function renderElement(container, element, place = `beforeend`) {
-  if (place === `beforeend`) {
-    container.append(element);
-    return;
-  }
-  if (place === `afterbegin`) {
-    container.prepend(element);
-    return;
+  switch (place) {
+    case (`beforeend`):
+      container.append(element);
+      break;
+    case (`afterbegin`):
+      container.prepend(element);
+      break;
   }
 }
